producer/pyproject.toml
[project]
name = "producer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Martijn Beenker", email = "Martijn.Beenker@gmail.com" }
]
requires-python = "~=3.11.0"
dependencies = [
    "pyspark~=3.5.0",
    "delta-spark~=3.3.0",
    "great-expectations[spark]~=1.4.0",
    "numpy==1.24.0",
    "pandas>=2.1.1",
]

[project.scripts]
producer = "producer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "faker>=37.1.0",
]
